добавил путь к ls
