+++ src/js/mobile-menu.js
@@ -1,0 +1,14 @@
+(() => {
+  const refs = {
+    openMenuBtn: document.querySelector('[data-menu-open]'),
+    closeMenuBtn: document.querySelector('[data-menu-close]'),
+    menu: document.querySelector('[data-menu]'),
+  };
+
+  refs.openMenuBtn.addEventListener('click', toggleModal);
+  refs.closeMenuBtn.addEventListener('click', toggleModal);
+
+  function toggleModal() {
+    refs.menu.classList.toggle('is-open');
+  }
+})();
