+++ src/partials/header.html
@@ -32,8 +32,9 @@
           <use href="../src/images/sprite.svg#icon-Burger"></use>
         </svg>
       </button>
-      <button class="mobile-btn hidden" type="button" data-modal-open>
-        Buy now >
+      <!-- <button class="mobile-btn hidden" type="button" data-modal-open> -->
+      <button class="button-arrow hidden" type="button" data-modal-open>
+        Buy now
       </button>
     </div>
   </div>
@@ -71,8 +72,8 @@
       </ul>
 
       <div class="mobile-burg-btn">
-        <button class="mobile-btn" type="button" data-modal-open>
-          Buy now >
+        <button class="button-arrow" type="button" data-modal-open>
+          Buy now
         </button>
       </div>
     </div>
